refactor: GitAPI로 Blob파일 가져오는 클래스들을 삭제합니다.
rate-limit 떄문에 GitAPI로 Blob파일을 가져오는건 실패해서 관련 클래스들을 제거합니다.

rate-limit는 personal 토큰당 약 5000/hour입니다.
