+++ api/src/test/resources/application-test.yaml
@@ -1,6 +1,4 @@
 spring:
-  config:
-    import: optional:file:../.env[.properties]
   jpa:
     hibernate:
       ddl-auto: create-drop
@@ -8,13 +6,6 @@ spring:
       hibernate:
         format_sql: true
         show_sql: true
-
-github:
-  personal-token: ${GITHUB_PERSONAL_TOKEN}
-
-opensource:
-  ingest-mode: local # "local" | "github-api"
-
 jwt:
   secretKey: ${JWT_SECRET_KEY}
   access-token-expiration: 600 # 10 분
