--- core/src/main/java/com/opensourcereader/core/analysis/dto/GitTreeFileDetail.java
@@ -1,11 +1,0 @@
-package com.opensourcereader.core.analysis.dto;
-
-public record GitTreeFileDetail(
-    String sha,
-    String node_id,
-    String size,
-    String url,
-    String content
-) {
-
-}
